+++ lib/rubocop/cop/rspec/multiple_expectations.rb
@@ -50,19 +50,37 @@ module RuboCop
 
         MSG = 'Example has too many expectations [%{total}/%{max}]'.freeze
 
-        def_node_search :expect, '(send _ :expect ...)'
+        def_node_matcher :expect?, '(send _ :expect ...)'
+        def_node_matcher :aggregate_failures?, <<-PATTERN
+        (block (send _ :aggregate_failures ...) ...)
+        PATTERN
 
         def on_block(node)
-          return unless example?(node) && (expectations = expect(node))
+          return unless example?(node)
 
-          return if expectations.count <= max_expectations
+          expectations_count = to_enum(:find_expectation, node).count
 
-          self.max = expectations.count
+          return if expectations_count <= max_expectations
 
-          flag_example(node, expectation_count: expectations.count)
+          self.max = expectations_count
+
+          flag_example(node, expectation_count: expectations_count)
         end
 
         private
+
+        def find_expectation(node, &block)
+          return unless node.is_a?(Parser::AST::Node)
+
+          yield if expect?(node) || aggregate_failures?(node)
+
+          # do not search inside of aggregate_failures block
+          return if aggregate_failures?(node)
+
+          node.children.each do |child|
+            find_expectation(child, &block)
+          end
+        end
 
         def flag_example(node, expectation_count:)
           method, = *node